refactor (backend): move fetch_evaluations_by_resource to db_manager and added function to remove testsets by ids
